feat: structural updates
