+++ interface/userprompt.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class UserPrompt(BaseModel):
+    prompt: str
+    address: str
