+++ services/utility/getmacaddress.py
@@ -1,0 +1,4 @@
+def get_mac_address():
+    mac_num = hex(uuid.getnode()).replace('0x', '').upper()
+    mac = ':'.join(mac_num[i : i + 2] for i in range(0, 11, 2))
+    return mac
